requirements.txt
customtkinter
openai
reportlab
python-dotenv
pillow
requests
cryptography
